feat: reduce package size
